Removing some more unused code, moving some code to src/features that no longer belongs in src/hooks
