--- src/hooks/useSelectedValueToText.ts
@@ -1,25 +1,0 @@
-import { useLanguage } from 'src/hooks/useLanguage';
-import { useOptionList } from 'src/hooks/useOptionList';
-import type { IUseLanguage } from 'src/hooks/useLanguage';
-import type { IOption, ISelectionComponent } from 'src/layout/common.generated';
-
-/**
- * Utility function meant to convert a value for a selection component to a label/text used in Summary
- *
- * Expected to be called from:
- * @see FormComponent.useDisplayData
- */
-export function useSelectedValueToText(component: ISelectionComponent, value: string) {
-  const langTools = useLanguage();
-  const optionList = useOptionList(component);
-  return getSelectedValueToText(value, langTools, optionList);
-}
-export function getSelectedValueToText(value: string, { langAsString }: IUseLanguage, optionList: IOption[]) {
-  const label = optionList.find((option) => option.value === value)?.label;
-
-  if (!label) {
-    return value;
-  }
-
-  return langAsString(label) || value;
-}
