+++ src/testUtils.tsx
@@ -205,26 +205,3 @@ export const mockComponentProps: IComponentProps & { id: string } = {
     throw new Error('Rendered mock legend, override this yourself');
   },
 };
-
-export const createStorageMock = (): Storage => {
-  let storage: Record<string, string> = {};
-  return {
-    setItem: (key, value) => {
-      storage[key] = value || '';
-    },
-    getItem: (key) => (key in storage ? storage[key] : null),
-    clear: () => {
-      storage = {};
-    },
-    removeItem: (key) => {
-      delete storage[key];
-    },
-    get length() {
-      return Object.keys(storage).length;
-    },
-    key: (i) => {
-      const keys = Object.keys(storage);
-      return keys[i] || null;
-    },
-  };
-};
