+++ src/utils/layout/index.tsx
@@ -1,55 +1,6 @@
 import { LayoutStyle } from 'src/layout/common.generated';
-import type { CompOrGroupExternal, ILayouts } from 'src/layout/layout';
 import type { ILayoutSet, ILayoutSets } from 'src/types';
 import type { IInstance } from 'src/types/shared';
-
-/**
- * @deprecated
- * @see useExprContext
- * @see useResolvedNode
- * @see ResolvedNodesSelector
- */
-export function getLayoutComponentById(id: string, layouts: ILayouts | null): CompOrGroupExternal | undefined {
-  if (!layouts) {
-    return undefined;
-  }
-
-  let component: CompOrGroupExternal | undefined;
-  Object.keys(layouts).forEach((layoutId) => {
-    if (!component) {
-      component = layouts[layoutId]?.find((element) => {
-        // Check against provided id, with potential -{index} postfix.
-        const match = matchLayoutComponent(id, element.id);
-        return match && match.length > 0;
-      });
-    }
-  });
-
-  return component;
-}
-
-/**
- * @deprecated
- * @see useExprContext
- * @see useResolvedNode
- * @see ResolvedNodesSelector
- */
-export function getLayoutIdForComponent(id: string, layouts: ILayouts): string | undefined {
-  let foundLayout: string | undefined;
-  Object.keys(layouts).forEach((layoutId) => {
-    if (!foundLayout) {
-      const component = layouts[layoutId]?.find((element) => {
-        // Check against provided id, with potential -{index} postfix.
-        const match = matchLayoutComponent(id, element.id);
-        return match && match.length > 0;
-      });
-      if (component) {
-        foundLayout = layoutId;
-      }
-    }
-  });
-  return foundLayout;
-}
 
 /**
  * Check if provided id matches component id.
